# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "酒店预订管理系统 API — 支持用户注册、酒店搜索、客房预订、订单管理、评价与数据统计",
        "title": "Hotel Booking API",
        "contact": {
            "name": "上海电力大学",
            "url": "https://www.shiep.edu.cn"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
...
        "/reviews": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "分页查询全部评价列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reviews"
                ],
                "summary": "查询全部评价",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/backend_model_schema.Review"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建新评价记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reviews"
                ],
                "summary": "创建评价",
                "parameters": [
                    {
                        "description": "评价信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend_model_schema.Review"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_model_schema.Review"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数无效",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/reviews/by-hotel": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据酒店 ID 分页查询该酒店的全部评价",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reviews"
                ],
                "summary": "按酒店查询评价",
                "parameters": [
                    {
                        "type": "string",
                        "description": "酒店 ID (UUID)",
                        "name": "hotelID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/backend_model_schema.Review"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/reviews/by-user": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据用户 ID 分页查询该用户的全部评价",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reviews"
                ],
                "summary": "按用户查询评价",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户 ID (UUID)",
                        "name": "userID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/backend_model_schema.Review"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/reviews/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据 UUID 查询单个评价信息（含关联用户、酒店、订单）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reviews"
                ],
                "summary": "查询评价详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "评价 ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_model_schema.Review"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "无效的评价 ID",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "404": {
                        "description": "评价不存在",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据 UUID 更新评价的评分和内容",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reviews"
                ],
                "summary": "更新评价",
                "parameters": [
                    {
                        "type": "string",
                        "description": "评价 ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "评分和内容",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.updateReviewInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_model_schema.Review"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据 UUID 硬删除评价",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reviews"
                ],
                "summary": "删除评价",
                "parameters": [
                    {
                        "type": "string",
                        "description": "评价 ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "backend_model_schema.Hotel": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "createAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "hotelName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend_model_schema.HotelImage"
                    }
                },
                "region": {
                    "$ref": "#/definitions/backend_model_schema.Region"
                },
                "regionID": {
                    "type": "integer"
                },
                "starLevel": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "telephone": {
                    "type": "string"
                },
                "updateAt": {
                    "type": "string"
                }
            }
        },
        "backend_model_schema.HotelImage": {
            "type": "object",
            "properties": {
                "hotel": {
                    "$ref": "#/definitions/backend_model_schema.Hotel"
                },
                "hotelID": {
                    "type": "string"
                },
                "imageURL": {
                    "type": "string"
                }
            }
        },
        "backend_model_schema.Order": {
            "type": "object",
            "properties": {
                "actualPrice": {
                    "type": "number"
                },
                "checkInDate": {
                    "type": "string"
                },
                "checkOutDate": {
                    "type": "string"
                },
                "createAt": {
                    "type": "string"
                },
                "discount": {
                    "type": "number"
                },
                "guests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend_model_schema.OrderGuest"
                    }
                },
                "id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "room": {
                    "$ref": "#/definitions/backend_model_schema.Room"
                },
                "roomID": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/backend_model_schema.OrderStatus"
                },
                "totalPrice": {
                    "type": "number"
                },
                "updateAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/backend_model_schema.User"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "backend_model_schema.OrderGuest": {
            "type": "object",
            "properties": {
                "idcard": {
                    "type": "string"
                },
                "order": {
                    "$ref": "#/definitions/backend_model_schema.Order"
                },
                "orderID": {
                    "type": "string"
                },
                "person": {
                    "$ref": "#/definitions/backend_model_schema.Person"
                }
            }
        },
...
        "backend_model_schema.Region": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "parent": {
                    "$ref": "#/definitions/backend_model_schema.Region"
                },
                "parentsID": {
                    "type": "integer"
                },
                "regionName": {
                    "type": "string"
                }
            }
        },
        "backend_model_schema.Review": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "createAt": {
                    "type": "string"
                },
                "hotel": {
                    "$ref": "#/definitions/backend_model_schema.Hotel"
                },
                "hotelID": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "order": {
                    "$ref": "#/definitions/backend_model_schema.Order"
                },
                "orderID": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                },
                "updateAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/backend_model_schema.User"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "backend_model_schema.Room": {
            "type": "object",
            "properties": {
                "availableQuantity": {
                    "type": "integer"
                },
                "createAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "facilities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend_model_schema.RoomFacility"
                    }
                },
                "hotel": {
                    "$ref": "#/definitions/backend_model_schema.Hotel"
                },
                "hotelID": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend_model_schema.RoomImage"
                    }
                },
                "price": {
                    "type": "number"
                },
                "status": {
                    "type": "integer"
                },
                "totalQuantity": {
                    "type": "integer"
                },
                "typeName": {
                    "type": "string"
                },
                "updateAt": {
                    "type": "string"
                },
                "weekendPrice": {
                    "type": "number"
                }
            }
        },
...
        "backend_model_schema.User": {
            "type": "object",
            "properties": {
                "createAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "idcard": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "points": {
                    "type": "integer"
                },
                "realName": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/backend_model_schema.UserRole"
                },
                "status": {
                    "type": "integer"
                },
                "updateAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "vipLevel": {
                    "$ref": "#/definitions/backend_model_schema.VipLevel"
                },
                "vipLevelID": {
                    "type": "integer"
                }
            }
        },
        "backend_model_schema.UserRole": {
            "type": "string",
            "enum": [
                "customer",
                "vip",
                "hotel_manager",
                "admin"
            ],
            "x-enum-varnames": [
                "RoleCustomer",
                "RoleVIP",
                "RoleHotelManager",
                "RoleAdmin"
            ]
        },
...
        "handler.updateReviewInput": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                }
            }
        },
        "model.ErrorObj": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "model.PaginationObj": {
            "type": "object",
            "properties": {
                "currentPage": {
                    "type": "integer"
                },
                "hasNext": {
                    "type": "boolean"
                },
                "hasPrev": {
                    "type": "boolean"
                },
                "itemsPerPage": {
                    "type": "integer"
                },
                "totalItems": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
        "model.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/model.ErrorObj"
                },
                "message": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/model.PaginationObj"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer Token，格式 \"Bearer \u003ctoken\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}